% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoid.R
\name{sim_sinusoid}
\alias{sim_sinusoid}
\title{Simulate Sinusoidal Wave to Data Frame}
\usage{
sim_sinusoid(
  f = "sin",
  n = 1,
  lambda = NULL,
  amp = 1,
  xmin = 0,
  xmax = 2 * pi,
  by = 0.05
)
}
\arguments{
\item{f}{(character or symbol) Function to simulate data from. By default, use \code{sin()} waves. Any trigonometric functions can be provided.}

\item{n}{(numeric) Number of cycle to simulate}

\item{lambda}{(numeric) Wave length in x-axis, if provided, will override \code{n}}

\item{amp}{(numeric) Amplitude of wave}

\item{xmin}{(numeric) Lower boundary of \code{x} in the output data}

\item{xmax}{(numeric) Upper boundary of \code{x} in the output data}

\item{by}{(numeric) Number to increment \code{x} by}
}
\value{
A data.frame with column: \code{x} and \code{y}
}
\description{
This function simulates sinusoidal wave data in a data frame format for further plotting. User can specify which trigonometric functions to simulate data from.
You can input parameters such as number of cycle to simulate (\code{n}), wave length, amplitude, and ranges of data to simulate.
}
\examples{
# Sinwaves with 2 cycle, amplitude = 3
df_sin1 <- sim_sinusoid("sin", n = 2, amp = 3)
plot(df_sin1, type = "l")
}
