% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoid.R
\name{sim_sinusoid_lm}
\alias{sim_sinusoid_lm}
\title{Simulate Sinusoidal Wave with Linear Model to Data Frame}
\usage{
sim_sinusoid_lm(
  f = "sin",
  n = 1,
  lambda = NULL,
  amp = 1,
  xmin = 0,
  xmax = 2 * pi,
  by = 0.05,
  slope = 0,
  y_intercept = 0,
  epsilon_sd = NULL
)
}
\arguments{
\item{f}{(character or symbol) Function to simulate data from. By default, use \code{sin()} waves. Any trigonometric functions can be provided.}

\item{n}{(numeric) Number of cycle to simulate}

\item{lambda}{(numeric) Wave length in x-axis, if provided, will override \code{n}}

\item{amp}{(numeric) Amplitude of wave}

\item{xmin}{(numeric) Lower boundary of \code{x} in the output data}

\item{xmax}{(numeric) Upper boundary of \code{x} in the output data}

\item{by}{(numeric) Number to increment \code{x} by}

\item{slope}{(numeric) Slope of the linear model to add to the waveform}

\item{y_intercept}{(numeric) Y-intercept of the linear model to add to the waveform}

\item{epsilon_sd}{If provided as numeric, it is a standard deviation of an error term sampled from Gaussian distribution with \code{mean = 0}}
}
\value{
A data.frame with column: \code{x} and \code{y}
}
\description{
Similar to \code{\link[=sim_sinusoid]{sim_sinusoid()}}, This function simulates sinusoidal wave, also, plus "linear model" to a data in a data frame format for further plotting.
You can input parameters such as number of cycle to simulate (\code{n}), wave length, amplitude, and ranges of data to simulate.
To add linear trend to the existing waves, you can specify arguments to adjust slope, y-intercept, and random error term.
}
\examples{
## Sine waves with upward linear trend
df_sin2 <- sim_sinusoid_lm(n = 5, slope = 1, epsilon_sd = 0.05)
plot(df_sin2, type = "l")
}
